feat(5-1): change structure of offers folder
