+++ src/store/nearby/selectors.js
@@ -1,7 +1,7 @@
 import {createSelector} from "reselect";
 
 import {MAX_NEARBY_OFFERS_COUNT} from "../../utils/constants.js";
-import {getCurrentOffers} from "../offers/selector.js";
+import {getCurrentOffers} from "../offers/selectors.js";
 import {getCurrentOffer} from "../cities/selectors.js";
 
 export const getNearby = createSelector(
