+++ src/store/offers/selectors.js
